fio_visualizer.py: add support to use against a remove server
fio lets you run fio on a remote server, and then connnect as a client
out to the server. You do this by issuing the following the the server:

  fio --server

Then on the localhost you'd run:

  fio --client=<hostname>

We can easily add support for using fio_visualizer by enabling the
user to enter an optional server. fio uses the nomenclature of "client"
for this, so we stick to that, and add a "Client" field for it. If
this is left empty we ignore it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>