realtime_back.py: fix use of fio eta argument
fio commit e592a06b36999 ("Add ETA output control and interactivity check")
added support for the --eta=[always|never] arguments. Since its addition
the algorithm has not changed: if you do not pass 'always' or 'never',
then the default is used. The default eventually was modified to be
named as 'auto', and this will print an ETA when the requested output is
a TTY.

As such, eta=0 really was forcing fio to use eta=auto. Fix this and use
eta=never. The realtime_back.py realtime parser only cares for the
output of the minimal terse output, so provide it only that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>