realtime_back.py: provide backward compatibility for fio versions
Using fio --minimal implies a set of loose defaults which one hopes
upstream fio won't break. For instance, the default terse version ouput
format used on the latest development version of fio is still version 3,
however this may change in the future.

Instead of relying on asking users to use a specific version of fio, we
can later check for minimum version compatibility, and use options which
are very specific. This also makes it clearer what the current
requirements are for fiovisualizer. We thus replace use of --minimal
with: --output-format=terse and --terse-version=3.

This should not break compatibility with any prior use given that:

  * When `--minimal` is used it is equivalent to using `--output-format=terse`.

  * The `--output-format` parameter was added via fio commit f3afa57e3 ("Add
    --output-format command line option") merged since fio-2.0.10.

  * fiovisualizer started supporting fio since fio-2.1.14. The release of
    fio-2.1.14 uses by default the terse output format version 3.
    fiovisualizer's parser only currently supports terse output format
    version 3.

  * The `--terse-version` parameter was added to fio via commit f57a9c59e36
    ("Add terse version output format command line parameter"), this
    commit was merged on the fio-1.58 release. This fio-1.58 release
    however used the terse version format 2. Development for the terse
    version format 3 started since fio commit 312b4af22018a ("Add IOPS
    to terse output") supported since fio-1.99.6.

  * The current default terse output format on the latest fio release is
    terse version format 3.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>