realtime_back.py: be more relaxed about output
If you use fiovisualizer to connect to a remote server the output
changes, and there are a slew of reasons why but the most important
one is that fio's terse fidelity is currently broken -- we get
different output if you try to get the terse output locally than if
the server is remote.

I've sent a patch to address this, but it may be the approach desired
by upstream fio. Even if this patch is accepted if you run --debug=net
the client will spit out network log data, for instance, and right now
fiovisualizer will barf at that.

Additionally as fio advances in features it may often happen, although
not intentionally, that more output is provided than expected. We
currently bake the required terse format column numbers to max 129
columns by forcing an array access to 130, which would throw an
exception error of IndexError().

While the number of columns still matches upstream behaviour this is
rather fragile.  There are many other reaons to not have this index
which is outside of the control of a client and debug is just one of the
reasons.

To address these issues:

  o check for a requied startswith "3;" prior to a split to ensure
    that the terse output format version is 3
  o simply ignore any other additional lines, and allow for flexibility
    if upstream fio terse version format v3 grows to more than
    129 columns

We can later expand our sanity checks on versioning and format for both
client / server, but this should be done separately.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>