README.md: recommend gfio
gfio is gfio own graphical solution and it already has proper
remote server / client support, and it is implements all sorts
of graphing solutions. Its architecture is more in line with what
a proper graphing tool should be for fio, and it has upstream
support.

Focus / love should go towards gfio then. If the maintainer of
fiovisualizer agrees, perahps this project should be declared as
deprecated.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>